+++ liquid/builtin/tags/include_tag.py
@@ -1,4 +1,5 @@
 """Parse tree node and tag definition for the built in "include" tag."""
+
 import sys
 from typing import Dict
 from typing import List
@@ -20,7 +21,6 @@ from liquid.expressions.common import parse_string_or_identifier
 from liquid.expressions.common import parse_unchained_identifier
 from liquid.expressions.filtered.parse import parse_obj
 from liquid.expressions.include.lex import tokenize
-from liquid.parse import expect
 from liquid.stream import TokenStream
 from liquid.tag import Tag
 from liquid.token import TOKEN_AS
@@ -192,7 +192,7 @@ class IncludeTag(Tag):
     def parse(self, stream: TokenStream) -> Node:
         """Read an IncludeNode from the given stream of tokens."""
         tok = next(stream)
-        expect(stream, TOKEN_EXPRESSION)
+        stream.expect(TOKEN_EXPRESSION)
         expr_stream = ExprTokenStream(
             tokenize(
                 stream.current.value,