+++ liquid/builtin/tags/increment_tag.py
@@ -11,7 +11,6 @@ from liquid.context import Context
 from liquid.expressions import TokenStream as ExprTokenStream
 from liquid.expressions.common import parse_unchained_identifier
 from liquid.expressions.filtered.lex import tokenize
-from liquid.parse import expect
 from liquid.stream import TokenStream
 from liquid.tag import Tag
 from liquid.token import TOKEN_EXPRESSION
@@ -51,10 +50,10 @@ class IncrementTag(Tag):
     node_class = IncrementNode
 
     def parse(self, stream: TokenStream) -> IncrementNode:
-        expect(stream, TOKEN_TAG, value=TAG_INCREMENT)
+        stream.expect(TOKEN_TAG, value=TAG_INCREMENT)
         tok = stream.current
         stream.next_token()
-        expect(stream, TOKEN_EXPRESSION)
+        stream.expect(TOKEN_EXPRESSION)
         return self.node_class(
             tok=tok,
             identifier=str(