+++ liquid/extra/tags/if_not.py
@@ -6,7 +6,6 @@ logical `not` operator and grouping terms with parentheses.
 
 from liquid.builtin.tags.if_tag import IfTag
 from liquid.expression import Expression
-from liquid.parse import expect
 from liquid.stream import TokenStream
 from liquid.token import TOKEN_EXPRESSION
 
@@ -18,7 +17,7 @@ class IfNotTag(IfTag):
 
     def parse_expression(self, stream: TokenStream) -> Expression:
         """Pare a boolean expression from a stream of tokens."""
-        expect(stream, TOKEN_EXPRESSION)
+        stream.expect(TOKEN_EXPRESSION)
         return self.env.parse_boolean_expression_value_with_parens(
             stream.current.value, stream.current.linenum
         )